feat: really early sketch of isolated instances
very very unfinished, pretty cool so far though
