+++ src/FirestormFiles.js
@@ -1,0 +1,74 @@
+const axios = require("axios").default;
+const { __extract_data } = require("./utils.js");
+
+/** Firestorm file manager */
+module.exports = class FirestormFiles {
+	/** Root Firestorm instance */
+	instance;
+
+	/**
+	 * Create a new Firestorm file manager based on a root instance
+	 * @param {ReturnType<import("./index.js").createFirestorm>} instance
+	 */
+	constructor(instance) {
+		this.instance = instance;
+	}
+
+	/** @ignore */
+	get __file_address() {
+		if (!this.instance.address)
+			throw new Error(`Address for Firestorm instance "${this.instance.name}" was not configured`);
+		return `${this.instance.address}files.php`;
+	}
+
+	/**
+	 * Get a file by its path
+	 * @memberof firestorm.files
+	 * @template T - Type of file content
+	 * @param {string} path - The wanted file path
+	 * @returns {Promise<T>} File contents
+	 */
+	get(path) {
+		return __extract_data(
+			axios.get(this.__file_address, {
+				params: {
+					path,
+				},
+			}),
+		);
+	}
+
+	/**
+	 * Upload a file
+	 * @memberof firestorm.files
+	 * @param {FormData} form - Form data with path, filename, and file
+	 * @returns {Promise<WriteConfirmation>} Write confirmation
+	 */
+	upload(form) {
+		form.append("token", this.instance.token);
+		return __extract_data(
+			axios.post(this.__file_address, form, {
+				headers: {
+					...form.getHeaders(),
+				},
+			}),
+		);
+	}
+
+	/**
+	 * Delete a file by its path
+	 * @memberof firestorm.files
+	 * @param {string} path - The file path to delete
+	 * @returns {Promise<WriteConfirmation>} Write confirmation
+	 */
+	delete(path) {
+		return __extract_data(
+			axios.delete(this.__file_address, {
+				data: {
+					path,
+					token: this.instance.token,
+				},
+			}),
+		);
+	}
+};
